requirements.txt
PyGithub
openai>=0.27.0
requests
